rename sponsor list files
