+++ src/containers/sponsor/list/all.spec.tsx
