+++ src/containers/sponsor/list/letter.graphql
